<Enhancement> 채팅 연결 시 팝업메시지와 함께 사이드바 색상변경, 페이지 이동시 default 값 #137
